整理了协议

+++ cat-net/src/main/java/com/cat/net/network/base/IProtocol.java
@@ -9,7 +9,7 @@ import com.google.protobuf.AbstractMessageLite.Builder;
  */
 public interface IProtocol {
 	
-	short protocol();
+	int protocol();
 	
 	default byte[] toBytes() {
 		return getBuilder().build().toByteArray();
